apache-airfllow/requirements.txt
apache-airflow==2.8.1
boto3==1.34.89
pandas==2.2.2
great_expectations==0.18.11
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2